requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
autopep8==2.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
django-silk==5.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
executing==2.0.1
Faker==24.11.0
gprof2dot==2022.7.29
idna==3.7
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
redis==5.0.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
toposort==1.10
traitlets==5.14.2
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
